devices: Rework test_file_parser
The tests in this module were overspecified and relied on the existence
of helper functions. Rework the tests to test through public interfaces
and remove the assumption that various helper functions exist. This
should allow us to make changes to this code more easily and have the
tests still be valid.